Renamed typeahead and typeaheadList to .ng1 and moved them to the inputs folder
